backend/requirements.txt
fastapi==0.111
uvicorn[standard]
sqlalchemy>=2.0
alembic
python-multipart
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
aiofiles
pydantic
gunicorn